Move contact-related commands into package contact
