+++ src/test/java/seedu/address/logic/parser/DeleteCommandParserTest.java
@@ -7,7 +7,7 @@ import static seedu.address.testutil.TypicalIndexes.INDEX_FIRST_PERSON;
 
 import org.junit.jupiter.api.Test;
 
-import seedu.address.logic.commands.DeleteCommand;
+import seedu.address.logic.commands.contact.DeleteCommand;
 
 /**
  * As we are only doing white-box testing, our test cases do not cover path variations
